% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_cal_log.R
\name{ag_cal_log}
\alias{ag_cal_log}
\title{Calibration Finder for specific ActiGraph devices using serial number}
\usage{
ag_cal_log(filepath)
}
\arguments{
\item{filepath}{The filepath where the noldus file exists}
}
\value{
Returns the recommended calibration offsets and scaling factor that should be applied to raw accelerometer data
}
\description{
\code{ag_cal_log} Will return the serial number of the selected file with the recommended calibration settings specific to that device
}
\examples{
acc_calibration_settings = ag_cal_log(filepath)

}
