% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag.epochr.R
\name{ag_epochr}
\alias{ag_epochr}
\title{Re-integreates ActiGraph data into whatever epoch you need, producing the same results that ActiLife does.}
\usage{
ag_epochr(ag_data_1sec, epoch = 60)
}
\arguments{
\item{epoch}{The new epoch length that you want your data to be in.}

\item{ag_data_1_sec}{The ActiGraph data in the original epoch you have.}
}
\value{
Returns the AcrtiGraph data in the new epoch.
}
\description{
\code{read.ag} Will return ActiGrpah data integreated into the epoch that you desire.
}
\examples{
actigraph_data_epoch = ag_epochr(ag_data_1sec = ag_data_1sec_hip, epoch = 60)
}
