% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO.info.R
\name{DO.info}
\alias{DO.info}
\title{Combines direct observation data with trimmed one-second count data from the accelerometer}
\usage{
DO.info(ag_data_1sec, noldus_expanded)
}
\arguments{
\item{ag_data_1sec}{The one-second count accelerometer data exported from trim.ag}

\item{noldus_expanded}{The Noldus data created from trim.noldus}
}
\value{
Returns the original 1-second accelerometer data with appended behaviors and activities.
}
\description{
\code{DO.info} Extracts the behavior and activity (Modifier 2) info from trim.noldus and appends observations to accelerometer output
}
\examples{
session_data = DO.info(Session_AG_1sec_hipdata, noldus_expanded)

}
