% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ag.R
\name{read.ag}
\alias{read.ag}
\title{Reads in the Noldus Observer XT output}
\usage{
read.ag(filepath, calibration_file)
}
\arguments{
\item{filepath}{The filepath where the ActiGraph file exists.}

\item{calibration_file}{The filepath for calibration data.}
}
\value{
Returns the AcrtiGraph data.
}
\description{
\code{read.ag} Will return ActiGrpah data; can handle count or raw data.
}
\examples{
actigraph_data = read.ag(filepath)
}
