% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.noldus.R
\name{read.noldus}
\alias{read.noldus}
\title{Reads in the Noldus Observer XT output}
\usage{
read.noldus(filepath)
}
\arguments{
\item{filepath}{The filepath where the noldus file exists}
}
\value{
Returns the Noldus excel file.
}
\description{
\code{read.noldus} Will return the Noldus excel file exported from Observer XT
}
\examples{
noldus_data = read.noldus(filepath)

}
