% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sed_cutoffs.R
\name{sed_cutoffs}
\alias{sed_cutoffs}
\title{Determines sedentary periods using various methods}
\usage{
sed_cutoffs(ag_data_1sec, ag_data_1sec_wrist, ag_data_raw,
  subj_label = participant, cpm100 = F, cpm150 = F, cpm200vm = F,
  soj1x = F, soj3x = F, sedsphere = F, jstaud2015 = F, wcpm1853vm = F,
  wcp15s376vm = F, enmo44.8 = F, VMcorrG_mod_15s = 489)
}
\arguments{
\item{ag_data_1sec}{Trimmed hip accelerometer data from trim.ag}

\item{ag_data_1sec_wrist}{Trimmed wrist accelerometer data from trim.ag}

\item{ag_data_raw}{Trimmed raw accelerometer data from trim.ag}

\item{subj_label}{Subject label identifier that will be appended as the first column in the returned data}

\item{cpm100}{Apply 100 counts/minute cutpoint to count data (Matthews et al. 2008)}

\item{cpm150}{Apply 150 counts/minute cutpoint to count data (Kozey-Keadle et al. 2011)}

\item{cpm200vm}{Apply 200 counts/minute using vector magnitude cutpoint to count data (Aguilar-Farias et al. 2014)}

\item{soj1x}{Apply Sojourn 1x algorithm to count data (Lyden et al. 2014)}

\item{soj3x}{Apply Sojourn 3x algorithm to count data (Lyden et al. 2014)}

\item{sedsphere}{Apply the Sedentary Sphere to raw wrist data (Rowlands et al. 2016)}

\item{jstaud2015}{Apply Decision Tree and Random Forest machine learning algorithms to raw wrist data (Staudenmayer et al. 2015)}

\item{wcpm1853vm}{Apply 1853 counts/minute using vector magnitude cutpoint to count data (Koster et al. 2016)}

\item{enmo44.8}{Apply 44.8 mg/minute using raw wrist data from read.ag or trim.ag (Hildebrand et al. 2017)}

\item{VMcorrG_mod_15s}{Vector magnitude corrected for gravity cutpoint used by Sedentary Sphere to determine activity periods, default is 489 mg}

\item{wcpm15s376vm}{Apply 376 counts/15 seconds using vector magnitude cutpoint to count data (Koster et al. 2016)}
}
\value{
Returns the 1-second accelerometer file with columns appended associated with the various methods to determine sedentary periods
}
\description{
\code{sed_cutoffs} Will return the 1-second accelerometer data with appended columns corresponding to the various methods indicating sedentary periods
}
\examples{
sed_data = sed_cutoffs(Session_AG_hip_1sec_data, Session_AG_wrist_1sec_data,Session_AG_wrist_raw_data, cpm100 = T)

}
