% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.ag.R
\name{trim.ag}
\alias{trim.ag}
\title{Trims ActiGraph data to give you data during specific intervals that you want to analyize}
\usage{
trim.ag(ag_data, vid_times, time_index, raw = F)
}
\arguments{
\item{ag_data}{The data that results from running read.ag}

\item{vid_times}{A spreadhseet that has the video start and stop times for each participant}

\item{time_index}{The index start and stop times of the particpant you want to analyize, from the origrinal spreadsheet}

\item{raw}{Indicate whether the ag_data file is raw or count data}
}
\value{
Returns the AcrtiGraph data trimmed to the specific time that you want to analyize.
}
\description{
\code{trim.ag} Will return ActiGrpah data integreated into the epoch that you desire.
}
\examples{
actigraph_data_epoch = ag_epochr(ag_data_1sec = ag_data_1sec_hip, epoch = 60)
}
