% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.noldus.R
\name{trim.noldus}
\alias{trim.noldus}
\title{Expands the Noldus data created by 'read.noldus'}
\usage{
trim.noldus(noldus_data, duration)
}
\arguments{
\item{noldus_data}{The Noldus data imported using read.noldus}

\item{duration}{The duration of observations of interest in seconds}
}
\value{
Returns the expanded Noldus data.
}
\description{
\code{trim.noldus} Will return the Noldus data expanded to 0.01 second data for aligning with accelerometer data
}
\examples{
session_noldus = trim.noldus(noldus_data)

}
